+++ src/structs/expression.js
@@ -12,6 +12,7 @@ class LimeExpression {
 
     // Properties
     this.name = args.hasOwnProperty('name') ? args.name : this.constructor.name.slice(12);
+    this.shorthand = args.hasOwnProperty('shorthand') ? args.shorthand : this.name;
     this.isSimple = args.hasOwnProperty('isSimple') && Types.isBoolean(args.isSimple) ? args.isSimple : true;
   }
 }