+++ src/expressions/matrix.js
@@ -6,7 +6,7 @@ class LimeExpressionMatrix extends LimeExpression {
   // Constructor
   constructor(lime, places) {
     // Super from expression class
-    super(lime, { name: 'matrix', shorthand: 'mat', isSimple: false });
+    super(lime, require('../../lib/expressions/matrix.json'));
 
     // Build expression
     if (!this.lime.identify('expression')(...places.flat())) {