+++ docs/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ### v0.4.0 (dev)
-- 
+- Converted `Compiler` from object to class
 
 ### v0.3.3 (Feb 5, 2022)
 - Extended `Equal` function to all expression types